requirements.txt
python-telegram-bot[job-queue]==21.9
pandas
requests
python-dotenv
openpyxl
python-calamine